requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pylint==3.2.2
pyflakes==3.2.0
bandit==1.7.8
openai==1.30.1
python-dotenv==1.0.1
astroid==3.2.2
